Add ImageUtils and fix some magic numbers
